feat(storage): implement comprehensive storage class and poolId validation. #1000664269
- Add validation for storage class in pool creation command
- Update error messages for invalid storage class inputs
- Enhance pool creation logic to ensure valid poolId usage
- Implement new methods for scanning inodes by poolId
- Modify inode update logic to include poolId validation
- Enhance logging for storage class and poolId related operations

+++ util/auditlog/auditlog.go
@@ -419,7 +419,7 @@ func (a *Audit) LogMigrationOp(clientAddr, volume, op, fullPath string, err erro
 	a.AddLog(entry)
 }
 
-func (a *Audit) LogLcNodeOp(op, vol, name, path string, pid, inode, size, leaseExpire uint64, hasMek bool, from, to uint32, latency int64, err error) {
+func (a *Audit) LogLcNodeOp(op, vol, name, path string, pid, inode, size, leaseExpire uint64, hasMek bool, from, to uint8, latency int64, err error) {
 	var errStr string
 	if err != nil {
 		errStr = err.Error()
@@ -553,7 +553,7 @@ func LogMigrationOp(clientAddr, volume, op, fullPath string, err error, latency 
 	gAdt.LogMigrationOp(clientAddr, volume, op, fullPath, err, latency, ino, from, to)
 }
 
-func LogLcNodeOp(op, vol, name, path string, pid, inode, size, leaseExpire uint64, hasMek bool, from, to uint32, latency int64, err error) {
+func LogLcNodeOp(op, vol, name, path string, pid, inode, size, leaseExpire uint64, hasMek bool, from, to uint8, latency int64, err error) {
 	gAdtMutex.RLock()
 	defer gAdtMutex.RUnlock()
 	if gAdt == nil {
